Finish exercise 02
Finished the exercise 02 "class-id-selectors". A comment I want to add
is that my solution is slightly off. One my class and id names are kinda
weird and dumb so I need to learn how to make better class and ID names
and I also forgot that if two or more elements uses a same rule, we can
"attach" the same class at it, example being the fourth and third
elements on said exercise, both share the 24px declaration so I could've
included both elements in one declaration. Yeaah much to learn still.