Finish cascade-fix lesson
As far as I read on the solution.css the way I solved this lesson isn't
commented on, so either I found another solution or my code is pure
spagetthi. I think its the second because I refused to change the
second ".para" position which yeah, looks like spagetthi code.