FInish lesson block and inline 02
This one was a doozy. I kinda forgot about the one liners for sizing
stuff, like "padding: 1, 1, 1, 1" so much so I forgot which one was it!
I found out that if you only put the first 2 numbers they get assigned
to top & bottom and left & right! My hunch is that it goes like this:
"padding: Top, Bottom, Right, Left", I need to check it out later.
Not only that I struggled with centering for a good while, like how does
"margin, 0 auto" centers the element? Apparently auto centers stuff? I
need to learn how it works.